+++ src/components/forms/form.module.css
@@ -1,0 +1,70 @@
+.box {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 0px;
+    position: absolute;
+    width: 480px;
+    height: 430px;
+    gap: 24px;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%)
+}
+
+.box_small {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-around;
+    width: 100%;
+}
+
+.box_order {
+    display: flex;
+    flex-direction: column;
+    padding: 0px;
+    position: absolute;
+    width: 480px;
+    height: 560px;
+    gap: 16px;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%)
+}
+
+.profile {
+    margin: 0;
+    display: flex;
+    flex-direction: row;
+    height: 100vh;
+    max-width: 1240px;
+
+}
+.profile_outlet {
+    margin: 016px;
+    width: 50%;
+}
+
+.navigation {
+   
+    padding: 0px;
+    width: 25%;
+   
+}
+
+.list {
+    padding: 0;
+    list-style-type: none;
+    display: grid;
+    grid-template-columns: repeat(4, 1fr);
+    gap: 20px;
+
+}
+
+.list_item {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    gap: 10px;
+}
