+++ src/components/forms/index.ts
@@ -1,0 +1,13 @@
+import { ForgotForm } from "./forgot-form";
+import { LoginForm } from "./login-form";
+import { RegisterForm } from "./register-form";
+import { ResetForm } from "./reset-form";
+import { UserForm } from "./user-form";
+
+export {
+    ForgotForm,
+    LoginForm,
+    RegisterForm,
+    ResetForm,
+    UserForm
+}
